refactor and leaving adding only the card related code
