--- src/components/Offer/Info/CompanyName/styles.module.scss
@@ -1,9 +1,0 @@
-@import "$styles/mixins/_subtitle.scss";
-@import "$styles/_variables.scss";
-
-.name {
-  @include subtitle;
-
-  color: $darkGrey;
-  font-weight: normal;
-}
