+++ src/components/Offer/Info/JobSpecs/index.tsx
@@ -1,11 +1,25 @@
 import React, { FunctionComponent } from "react";
+import AttachMoneyIcon from "@material-ui/icons/AttachMoney";
+import AccessTimeIcon from "@material-ui/icons/AccessTime";
+import { SpecItem } from "./SpecItem";
 
-import styles from "./styles.module.scss";
-
-const JobSpecs: FunctionComponent = ({children}) => (
-  <h2 className={styles.name}>
-  {children}
-  </h2>
+const JobSpecs: FunctionComponent<IJobSpecsProps> = ({salary, workload}) => (
+  <div>
+    <SpecItem item={workload.time} description={workload.description}>
+      <AccessTimeIcon fontSize="small" />
+    </SpecItem>
+    <SpecItem item={salary} >
+    <AttachMoneyIcon fontSize="small"/>
+    </SpecItem>
+  </div>
 );
+
+interface IJobSpecsProps {
+  salary: string;
+  workload: {
+    time: string;
+    description: string;
+  };
+}
 
 export { JobSpecs };
