--- src/components/Offer/Info/Title/styles.module.scss
@@ -1,8 +1,0 @@
-@import "$styles/mixins/_subtitle.scss";
-@import "$styles/_variables.scss";
-
-.title {
-  @include subtitle;
-
-  margin-bottom: $smallSpacing;
-}
