+++ src/components/Offer/Info/styles.module.scss
@@ -1,17 +1,18 @@
 @import "$styles/_variables.scss";
 
+.jobDescription {
+  font-weight: bold;
+}
+
 .container {
-  // display: flex;
-  // flex-grow: 2;
   flex-basis: 100%;
   margin: $smallSpacing;
   margin-left: $bigSpacing;
   flex-direction: column;
-  // background-color:green;
 }
 
 .separator {
-  border: .5px solid $grey;
+  border: .5px solid lighten($color: $grey, $amount: 30%);
 }
 
 .detailsContainer {
@@ -19,5 +20,9 @@
   margin: $bigSpacing;
   margin-left: $bigSpacing;
   margin-bottom: 0;
+  justify-content: space-around;
+}
+
+.firstColumn {
   flex-direction: column;
 }
