+++ src/pages/Applicant/Home/Feed/index.tsx
@@ -1,1 +1,1 @@
-export { FeedContainer as Feed } from "./container";
+export { Feed } from "./component";
