+++ src/styles/mixins/_subtitle.scss
@@ -1,5 +1,5 @@
 @mixin subtitle {
-  color: black;
+  color: $darkGrey;
   margin: 0;
   font-weight: bold;
   font-size: 22px;
